Build a npm package
- Create an npm folder in /dist/pkg
- Copy in the package.json file (we probably want to create this,
rather than copy it here)
- Zip up the contents of this folder

Where I’m using paths.distPkgNpm, it would be better to pass in the
package type, so this can be reused for other packages.

Also, I’m not sure it’s useful to keep the contents of dist/pkg/npm
once the zip has been produced so it might be worth adding a clean step
to remove assets and templates from this folder.