Add block variable for body attributes
I've inheritied a project which relies on GA event tagging being driven by attributes dynamically added to the body tag.
Would you accept a modification which allows for a block variable being added to the body tag. This should be a non-breaking change for anyone not using this block var.